ModArith: Refactor op-related things into separated file
